+++ src/app/pages/editar-empresa/editar-empresa.component.html
@@ -2,8 +2,8 @@
     <div class="p-col-10">
         <p-card class="p-shadow-10 ">
             <ng-template pTemplate="header">
-                <p-panel header="Formulário para Cadastro">
-                    <app-formulario-empresas [pessoaJuridica]=pessoaJuridica></app-formulario-empresas>
+                <p-panel header="Formulário para Editar">
+                    <app-formulario-empresas [pessoaJuridica] =pessoaJuridica></app-formulario-empresas>
                 </p-panel>
             </ng-template>
         </p-card>
