+++ src/app/pages/editar-empresa/editar-empresa.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { EditarEmpresaComponent } from './editar-empresa.component';
+
+describe('EditarEmpresaComponent', () => {
+  let component: EditarEmpresaComponent;
+  let fixture: ComponentFixture<EditarEmpresaComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ EditarEmpresaComponent ]
+    })
+    .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(EditarEmpresaComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
