+++ src/app/services/empresa.service.ts
@@ -29,6 +29,10 @@ export class EmpresaService {
     return this.http.get<PessoaJuridica[]>(`${API_CONFIG.baseUrl}/cadastro/matrizes`);
   }
 
+  getEmpresaId(id: string | null): Observable<PessoaJuridica> {
+    return this.http.get<PessoaJuridica>(`${API_CONFIG.baseUrl}/cadastro/${id}`);
+  }
+
   listarFiltroEmpresa(empresaFilter:PessoaJuridicaFilter) {
     return this.http.post<PessoaJuridica[]>(`${API_CONFIG.baseUrl}/cadastro/filtrar`,empresaFilter);
   }
